requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
click==7.1.2
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-gis==0.15
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
